+++ tests/testthat/test-errors-silent-stop.R
@@ -76,9 +76,6 @@ test_that("caesar - returns errors", {
 
 
   # shift isn't a string
-  expect_error(caesar(2,
-                      shift = 2),
-               "text must be a string or vector of strings!")
   expect_error(caesar(as.Date("2011-01-22"),
                       shift = 2),
                "text must be a string or vector of strings!")
@@ -90,24 +87,12 @@ test_that("caesar - returns errors", {
                       shift = 2,
                       decrypt = TRUE),
                "text must be a string or vector of strings!")
-  expect_error(caesar(2:10,
-                      shift = 2,
-                      decrypt = TRUE),
-               "text must be a string or vector of strings!")
 
   expect_error(caesar(TRUE,
                       shift = 2),
                "text must be a string or vector of strings!")
   expect_error(caesar(FALSE,
                       shift = 2),
-               "text must be a string or vector of strings!")
-  expect_error(caesar(NULL,
-                      shift = 2,
-                      decrypt = TRUE),
-               "text must be a string or vector of strings!")
-  expect_error(caesar(NA,
-                      shift = 2,
-                      decrypt = TRUE),
                "text must be a string or vector of strings!")
 
 
@@ -177,9 +162,6 @@ test_that("seed - returns errors", {
 
 
   # seed isn't a single string
-  expect_error(seed_cipher(2,
-                           seed = 2),
-               "text must be a string or vector of strings!")
   expect_error(seed_cipher(as.Date("2011-01-22"),
                            seed = 2),
                "text must be a string or vector of strings!")
@@ -188,10 +170,6 @@ test_that("seed - returns errors", {
                            decrypt = TRUE),
                "text must be a string or vector of strings!")
 
-  expect_error(seed_cipher(2:10,
-                           seed = 2,
-                           decrypt = TRUE),
-               "text must be a string or vector of strings!")
 
   expect_error(seed_cipher(TRUE,
                            seed = 2),