+++ NEWS.md
@@ -1,3 +1,8 @@
+# caesar 1.1.1
+
+* Fix bug where an input is NA or NULL causes an error. Now returns that input (NA if NA, NULL if NULL).
+* Can now encrypt numeric values. Works the same as with strings. 
+
 # caesar 1.1.0
 
 * Can now encrypt or decrpyt vector of strings (previously could only do a single string at a time).